+++ frontend/src/components/addToCartButton/addToCartButton.scss
@@ -1,0 +1,6 @@
+.add-to-cart-button {
+  background-color: #eda66e !important;
+  border: solid 1px black !important;
+  color: black !important;
+  float: right !important;
+}
