+++ frontend/src/components/product/product.scss
@@ -18,12 +18,7 @@
             }
         }
 
-        .btn-primary{
-            background-color: #eda66e;
-            border: solid 1px black;
-            color: black;
-            float: right;
-        }
+        
     
 
     }
